Use std::vector instead of AmiraMesh HxParamBundle to store nonmortar triangle renumbering
When used as a mapping implementation for a domain decomposition algorithm,
the triangles of the nonmortar (domain) surface need to know which triangles
of the original domain boundary they belong to.  This information used to
be stored in a HxParamBundle object.  This needed to be replace, because
it requires libamiramesh.