Bugfix: number of target segments wasn't computed correctly.
Back in the days when the array 'TargetSegments' was a plain int array,
the size of that array had to be divided by two to get the actual number
of segments.  Now that it's an array of std::array, this isn't necessary
anymore.  But I forgot to remove one of those /2.