Change class name from 'Parametrization' to 'PSurface', and make it a class template.
This should again make for a clearer naming.  The template parameters are the surface
dimension and the type used for coordinates.  While only 2d surfaces are currently
implemented, I may want to move some infrastructure from the 1d ContactMapping code
to a 1d specialication of PSurface.  As for the coordinate type:  for numerics
applications we need support for double precision, while for computer graphics
we want float.  Hence the type is a parameter now.  At this time, both types are
completely ignored, however.

In the library, the template is instantiated for dim=2 and type=float,double,
only.  If you need more you have to edit PSurface.cpp