Revert "psurface: [Compile fix] Use MC_isnan to compile on MacX"
This reverts commit cb6f03efc72f0ca91871a44abc8781cbeeaf0911.

The reason for this revert is that psurface is also used independently
of Amira as a stand alone program.  Hence, it must not depend on
mclib.  Thus, a different fix for MacX is needed.  However, since MacX
is not currently supported anyway, we can safely revert this change
without breaking compiling.