Also instantiate class Hdf5IO for 1d psurfaces
And add a few hacks that allow that to compile.

1d and 2d psurfaces are vastly different.  Syntactically they don't have
anything in common.  In particular, hdf5 IO will not work for 1d psurfaces.
The 1d version of hdf5 io needs to be in the library, though, for dune-grid
to link properly against psurface.  Therefore this patch adds a little
nasty hack to make the linking succeed.