Always check for MPI, even if not requested by configure options
If the hdf5 installation is compiled against MPI, then the test for hdf5
needs the MPI flags to find it.  Therefore, we need to run the test for
MPI even though we never actually use it.  Hence it should always run,
the user cannot be expected to turn it on explicitly.